RA-1386 moved existing validation class and interface to validation folder.
