+++ src/Esfa.Vacancy.Register.Api/Esfa.Vacancy.Register.Api.csproj
@@ -25,6 +25,7 @@
     <NuGetPackageImportStamp>
     </NuGetPackageImportStamp>
     <TargetFrameworkProfile />
+    <Use64BitIISExpress />
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -285,12 +286,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App_Start\AutoMapperConfig.cs" />
-    <Compile Include="App_Start\IValidationBadRequestBuilder.cs" />
     <Compile Include="App_Start\VacancyApiExceptionHandler.cs" />
     <Compile Include="App_Start\CustomErrorResult.cs" />
     <Compile Include="App_Start\ApiFilterConfig.cs" />
     <Compile Include="App_Start\StrictEnumConverter.cs" />
-    <Compile Include="App_Start\ValidationBadRequestBuilder.cs" />
     <Compile Include="Constants\RouteName.cs" />
     <Compile Include="Controllers\GlobalErrorController.cs" />
     <Compile Include="Controllers\GetTraineeshipVacancyController.cs" />
@@ -326,6 +325,8 @@
     <Compile Include="Orchestrators\SearchApprenticeshipVacanciesOrchestrator.cs" />
     <Compile Include="Orchestrators\GetApprenticeshipVacancyOrchestrator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Validation\IValidationBadRequestBuilder.cs" />
+    <Compile Include="Validation\ValidationBadRequestBuilder.cs" />
     <Compile Include="WebRole.cs" />
   </ItemGroup>
   <ItemGroup>
